Buisness logic implemented

+++ src/payment/comment_gen.py
@@ -3,7 +3,7 @@ import string
 import datetime
 
 
-def gen_comment(user_id: int, amount: int, k: int = 10):
+def gen_comment(user_id: int, amount: float, k: int = 10):
     random_str = "".join(
         random.choices(
             string.ascii_uppercase + string.digits + string.ascii_lowercase, k=k
